Fix #3: wrong usage of with-current-buffer
`with-output-to-string` will temporarily change variable
`standard-output` with value `"#<buffer *string-output*>"`.

and the code

```elisp
  (with-current-buffer
      standard-output
    ...)
```

will change the current buffer into buffer `standard-output`,
that's the buffer `"#<buffer  *string-output*>"`,
and use functions like `point-min`, `point-max`, `shell-command-on-region` is pointless,
because the current buffer is an empty buffer.

Check https://stackoverflow.com/questions/3576192 for details.